requirements.txt
streamlit==1.32.0
tensorflow>=2.16.1
pandas>=2.2.0
numpy<2.0.0
Pillow>=10.0.0
plotly>=5.18.0
reportlab>=4.0.0
opencv-python>=4.8.0
xlsxwriter>=3.1.0